requirements.txt
fastapi
uvicorn[standard]
authlib
httpx
python-dotenv
Pillow
numpy
scikit-image
psycopg2-binary
discord.py
pydantic
itsdangerous
python-multipart
